% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asko3c.R
\name{asko3c}
\alias{asko3c}
\title{asko3c}
\usage{
asko3c(data_list, parameters)
}
\arguments{
\item{data_list, }{list contain all data and metadata (DGEList, samples descriptions, contrast, design and annotations)}

\item{parameters, }{list that contains all arguments charged in Asko_start}
}
\value{
asko, list of data.frame contain condition, contrast and context information
}
\description{
Create contrast/condition/context file in format readable by Askomics Software.
}
\examples{
\dontrun{
    parameters <- Asko_start()
    data<-loadData(parameters)
    asko<-asko3c(data, parameters)
}

}
