% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askoR-package.R
\docType{package}
\name{askoR-package}
\alias{askoR}
\alias{askoR-package}
\title{askoR: Differential Expression Analysis using 'edgeR'}
\description{
Pipeline for the analysis of gene expression data, using 'edgeR' for differential analysis, 'topGO' for GOenrichment, and 'coseq' for gene clustering.
}
\author{
\strong{Maintainer}: Kévin Gazengel \email{kevin.gazengel@inrae.fr}

Authors:
\itemize{
  \item Fabrice Legeai \email{fabrice.legeai@inrae.fr}
  \item Susete Alves Carvalho \email{susete.alves-carvalho@inrae.fr}
}

}
\keyword{internal}
