% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskoStats.R
\name{AskoStats}
\alias{AskoStats}
\title{AskoStats}
\usage{
AskoStats(glm_test, fit, contrast, ASKOlist, dge, parameters)
}
\arguments{
\item{glm_test, }{tests for one or more coefficients in the linear model (likelihood ratio tests or empirical Bayes quasi-likelihood F-tests).}

\item{fit, }{fitted linear model object.}

\item{contrast, }{coefficient/contrast names tested.}

\item{ASKOlist, }{list of data.frame contain condition, contrast and context informations made by asko3c.}

\item{dge, }{large DGEList with normalized counts by GEnorm function.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none
}
\description{
Based on result contained in "glm_test":
\itemize{
\item print summary result of differential expression analysis
\item format all results in tabulate out file followed parameters given
\item plot heatmap of top differential expressed genes
}
Create one file by contrast, each file contains for each genes: fold-change and
log fold-change values, PValue, Expression, Significance, logCPM, LR, FDR and
significance value for each condition/context.
By default, LR and logCPM were not displayed, you can switch this parametres
to TRUE for display.
}
\examples{
\dontrun{
    AskoStats(glm_test, fit, contrast, ASKOlist, dge, parameters)
}

}
