% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustAndGO.R
\name{ClustAndGO}
\alias{ClustAndGO}
\title{ClustAndGO}
\usage{
ClustAndGO(asko_norm, resDEG, parameters, data, list = NULL, title = NULL)
}
\arguments{
\item{asko_norm}{large DGEList with normalized counts by GEnorm function.}

\item{resDEG}{data frame contains for each contrast the significance expression (1/0/-1) for all genes coming from DEanalysis function.}

\item{parameters}{list that contains all arguments charged in Asko_start.}

\item{data}{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations)}

\item{list}{gene list of interest if you want to apply ClustAndGO function on a specific gene list}

\item{title}{name of the gene list if you want to apply ClustAndGO function on a specific gene list}
}
\value{
data frame with clusters of each gene
}
\description{
Clusterize genes with same profile, proceed to GO-enrichment on clusters, search for contrasts enriched with genes of specific clusters, and identify intersections of DE list genes
\itemize{
\item Graphs of clusters (heatmap and boxplot)
\item Expression profiles in each cluster
\item GO enrichments in each cluster
\item Files with gene description of each significant enriched GO
\item Over-representation of genes of each cluster in each contrast
\item Intersections of DE list genes in each cluster
}
}
\note{
Remember to read the Wiki section in \url{https://github.com/askomics/askoR/wiki}
}
\examples{
\dontrun{
   clust<-ClustAndGO(asko_norm, resDEG, parameters, data)
   # OR
   clust<-ClustAndGO(asko_norm, resDEG, parameters, data, list, title)
}

}
