% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{DEanalysis}
\alias{DEanalysis}
\title{DEanalysis}
\usage{
DEanalysis(norm_GE, data_list, asko_list, parameters)
}
\arguments{
\item{norm_GE, }{large DGEList with normalized counts and data description.}

\item{data_list, }{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations).}

\item{asko_list, }{list of data.frame contain condition, contrast and context informations made by asko3c.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
SumMat, list (TestResults format class limma) contains for each contrast the significance expression (1/0/-1) for all gene.
}
\description{
Genewise statistical tests for a given coefficient or contrast, with edgeR method.
}
\examples{
\dontrun{
    sum_table<-DEanalysis(norm_GE, data_list, asko_list, parameters)
}

}
