% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEcorr.R
\name{GEcorr}
\alias{GEcorr}
\title{GEcorr}
\usage{
GEcorr(asko_norm, parameters)
}
\arguments{
\item{asko_norm, }{large DGEList with normalized counts by GEnorm function.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none
}
\description{
Plot some graphes to see data correlation :
\itemize{
\item heatmap sample correslation
\item MDS plots
\item hierarchical clustering
}
}
\examples{
\dontrun{
    GEcorr(asko_norm,parameters)
}

}
