% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEfilt.R
\name{GEfilt}
\alias{GEfilt}
\title{GEfilt}
\usage{
GEfilt(data_list, parameters)
}
\arguments{
\item{data_list, }{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations)}

\item{parameters, }{list that contains all arguments charged in Asko_start}
}
\value{
filtered_counts, large DGEList with filtered counts and data descriptions.
}
\description{
\itemize{
\item Filter genes according to cpm threshold value and replicated cpm option value.
\item Plot different graphes to explore data before and after filtering.
}
}
\examples{
\dontrun{
    filtered_counts<-GEfilt(data, parameters)
}

}
