% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneInfo_OnList.R
\name{GeneInfo_OnList}
\alias{GeneInfo_OnList}
\title{GeneInfo_OnList}
\usage{
GeneInfo_OnList(
  list,
  resDEG,
  data,
  parameters,
  title,
  clustering = NULL,
  conditions = NULL,
  contrasts = NULL
)
}
\arguments{
\item{list}{list contain all the genes you want to get information on}

\item{resDEG}{data frame contains for each contrast the significance expression (1/0/-1) for all genes coming from DEanalysis function.}

\item{data}{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations)}

\item{parameters}{list that contains all arguments charged in Asko_start.}

\item{title}{name of the gene list}

\item{clustering}{data frame with clusters of each gene produced by ClustAndGO function}

\item{conditions}{list of the conditions you want to see in graph and table}

\item{contrasts}{list of the conditions you want to see in graph and table}
}
\value{
none
}
\description{
Produce a heatmap of the expression of the genes included in the list with information on DE status of each gene and resume all informations on the genes in a table
\itemize{
\item Table with gene expression, DE status, clustering, and gene description
\item Heatmap of gene expression and DE status
}
}
\note{
Remember to read the Wiki section in \url{https://github.com/askomics/askoR/wiki}
}
\examples{
\dontrun{
   GeneInfo_OnList(list, resDEG, data, parameters, title)
   # OR
   GeneInfo_OnList(list, resDEG, data, parameters, title, clustering)
   # OR
   GeneInfo_OnList(list, resDEG, data, parameters, title, clustering, conditions)
   # OR
   GeneInfo_OnList(list, resDEG, data, parameters, title, clustering, conditions, contrasts)
}

}
