% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEnorm.R
\name{GEnorm}
\alias{GEnorm}
\title{GEnorm}
\usage{
GEnorm(filtered_GE, asko_list, data_list, parameters)
}
\arguments{
\item{filtered_GE, }{large DGEList with filtered counts by GEfilt function.}

\item{asko_list, }{list of data.frame contain condition, contrast and context informations made by asko3c.}

\item{data_list, }{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations)}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
norm_GE, large DGEList with normalized counts and data descriptions.
}
\description{
Normalize counts
\itemize{
\item Calculate normalization factors to scale the filtered library sizes.
\item Plot different graphes to explore data before and after normalization.
\item Optionally, write file with mean counts and normalized mean counts in Askomics format.
}
}
\examples{
\dontrun{
    norm_GE<-GEnorm(filtered_counts, asko, parameters)
}

}
