% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichment.R
\name{GOenrichment}
\alias{GOenrichment}
\title{GOenrichment}
\usage{
GOenrichment(resDEG, data_list, parameters)
}
\arguments{
\item{resDEG, }{data frame contains for each contrast the significance expression (1/0/-1) for all gene.}

\item{data_list, }{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations).}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none.
}
\description{
Perform GO enrichment analysis with topGO package.
This package provides tools for testing GO terms while accounting for
the topology of the GO graph. Different test statistics and different
methods for eliminating local similarities and dependencies between GO
terms can be implemented and applied.
}
\examples{
\dontrun{
   GOenrichment(resDEG, data_list, parameters)
}

}
