% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncludeNonDEgenes_InClustering.R
\name{IncludeNonDEgenes_InClustering}
\alias{IncludeNonDEgenes_InClustering}
\title{IncludeNonDEgenes_InClustering}
\usage{
IncludeNonDEgenes_InClustering(data, asko_norm, resDEG, parameters, clustering)
}
\arguments{
\item{data}{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations)}

\item{asko_norm}{large DGEList with normalized counts by GEnorm function.}

\item{resDEG}{data frame contains for each contrast the significance expression (1/0/-1) for all genes coming from DEanalysis function.}

\item{parameters}{list that contains all arguments charged in Asko_start.}

\item{clustering}{data frame with clusters of each gene produced by ClustAndGO function}
}
\value{
none
}
\description{
Add a cluster with the genes that are not DE in the analysis to the ClustAndGO analysis
\itemize{
\item Graphs of clusters (heatmap and boxplot) created through ClustAndGO function with an additionnal cluster représenting NON DE genes
\item Expression profile of NON DE genes
\item GO enrichments on NON DE genes
\item Files with gene description of each significant enriched GO
}
}
\note{
Remember to read the Wiki section in \url{https://github.com/askomics/askoR/wiki}
}
\examples{
\dontrun{
   IncludeNonDEgenes_InClustering(data, asko_norm, resDEG, parameters, clustering)
}

}
