% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(parameters)
}
\arguments{
\item{parameters, }{list that contains all arguments charged in Asko_start}
}
\value{
data, list contain all data and metadata (DGEList, samples descriptions, contrast, design and annotations)
}
\description{
Function to load :
\itemize{
\item Count data (one of this below):
\itemize{
\item count matrix : 1 file with all counts for each samples/conditions or multiple
\item list of files : 1 file of count per conditions, files names contained in sample file
}
\item Metatdata :
\itemize{
\item sample file : file describing the samples and the experimental design
\item contrast file : matrix which specifies which comparisons you would like to make between the samples
\item (optional) annotation file : functional/genomic annotation for each genes
\item (optional) GO terms annotations files : GO annotations for each genes
}
}
Three output directory will be create :
\itemize{
\item images : contains all the images created when running this pipeline with the exception of the upsetR and Venn graphs.
\item vennDiagram : contain all venn diagrams created by VD function
\item UpSetR_graphs : contain all upset graphs created by UpSetGraph function
\item Askomics : files compatible with Askomics Software
}
}
\examples{
\dontrun{
    parameters<-Asko_start()
    data<-loadData(parameters)
}

}
