% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormCountMean.R
\name{NormCountsMean}
\alias{NormCountsMean}
\title{NormCountsMean}
\usage{
NormCountsMean(glmfit, ASKOlist, context = NULL)
}
\arguments{
\item{glmfit, }{fitted linear model object.}

\item{ASKOlist, }{list of data.frame contain condition, contrast and context information made by asko3c.}

\item{context, }{coefficient/contrast tested.}
}
\value{
one of this:
\itemize{
\item matrixMean, matrix with mean counts,
\item meanValue for one context/Condition.
}
}
\description{
Calculation mean counts for two contrast or all matrix.
}
\examples{
\dontrun{
    # calculate mean counts in contrast contx1_vs_contx2
    mean1<-NormCountsMean(glmfit, ASKOlist, contx1)   # in the 1st context
    mean2<-NormCountsMean(glmfit, ASKOlist, contx2)   # in the 2nd context

    # for all conditions
    n_count<-NormCountsMean(glmfit, ASKOlist, context=NULL)
}

}
