% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expr.R
\name{plot_expr}
\alias{plot_expr}
\title{plot_expr}
\usage{
plot_expr(fit, normGE, resDE, contrast, tplot, parameters)
}
\arguments{
\item{fit, }{fitted linear model object.}

\item{normGE, }{large DGEList with normalized counts and data description.}

\item{resDE, }{vector containing integer values of -1 to represent down-regulated
genes, 0 for no differential expression, and 1 for up-regulated genes.}

\item{contrast, }{coefficient/contrast tested.}

\item{tplot, }{type of plot selected for display.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none.
}
\description{
Function to generate plots showing different aspects of differential expression results.
Two types of graphs can be created:
\itemize{
\item Mean-Difference Plot of Expression Data (aka MA plot).
\item Volcano plot for a specified coefficient/contrast of a linear model.
}
}
\examples{
\dontrun{
   plot_expr(fit, normGE, resDE, contrast, "MD", parameters)  # smear plot
   plot_expr(fit, normGE, resDE, contrast, "VO", parameters)  # volcano plot
}

}
