% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glimma.R
\name{plot_glimma}
\alias{plot_glimma}
\title{plot_glimma}
\usage{
plot_glimma(fit, normGE, resDE, contrast, tplot, parameters)
}
\arguments{
\item{fit, }{fitted linear model object.}

\item{normGE, }{large DGEList with normalized counts and data description.}

\item{resDE, }{vector containing integer values of -1 to represent down-regulated
genes, 0 for no differential expression, and 1 for up-regulated genes.}

\item{contrast, }{coefficient/contrast tested.}

\item{tplot, }{type of plot selected for display.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none.
}
\description{
Use Glimma package for interactive visualization of results from differential
expression analyses. Two types of graphs can be created:
\itemize{
\item Mean-Difference Plot of Expression Data (aka MA plot).
\item Volcano plot for a specified coefficient/contrast of a linear model.
}
}
\examples{
\dontrun{
   plot_glimma(fit, normGE, resDE, contrast, "MD", parameters)  # smear plot
   plot_glimma(fit, normGE, resDE, contrast, "VO", parameters)  # volcano plot
}

}
