% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSetGraph.R
\name{UpSetGraph}
\alias{UpSetGraph}
\title{UpSetGraph}
\usage{
UpSetGraph(resDEG, data_list, parameters)
}
\arguments{
\item{resDEG, }{data frame contains for each contrast the significance expression (1/0/-1) for all gene.}

\item{data_list, }{list contain all data and metadata (DGEList, samples descritions, contrast, design and annotations).}

\item{parameters, }{list that contains all arguments charged in Asko_start.}
}
\value{
none
}
\description{
Generate upsetR graphs.
}
\examples{
\dontrun{
   UpSetGraph(sumDEG, data_list, parameters)
}

}
