% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VD.R
\name{VD}
\alias{VD}
\title{VD}
\usage{
VD(resDEG, parameters, asko_list)
}
\arguments{
\item{resDEG, }{data frame contains for each contrast the significance expression (1/0/-1) for all gene.}

\item{parameters, }{list that contains all arguments charged in Asko_start.}

\item{asko_list, }{list of data.frame contain condition, contrast and context informations made by asko3c.}
}
\value{
none.
}
\description{
Plot Venn Diagram to compare different contrast
}
\examples{
\dontrun{
   VD(resDEG, parameters, asko_list)
}

}
