% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asko_start.R
\name{Asko_start}
\alias{Asko_start}
\title{Asko_start}
\usage{
Asko_start()
}
\value{
List of parameters that contains all arguments.
}
\description{
Initialize and Scans parameters from command line in a python-like style:
\itemize{
\item declare options, their flags, types, default values and help messages,
\item read the arguments passed to the R script and parse them according to what has been declared in step 1.
}
Parameters can be called by their names as declared in opt object.
}
\note{
All parameters were describe in Wiki section: \url{https://github.com/askomics/askoR/wiki}
}
\examples{
   parameters <- Asko_start()

   # set inputs files
   parameters$fileofcount = "CountsMatrix.txt"
   parameters$sample_file = "Samples_CountsMatrix.txt"
   parameters$contrast_file = "Contrasts.txt"

   # Set parameters threshold cpm
   parameters$threshold_cpm <- 0.5

   # And other ... askoR has many parameters.

}
