% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askoR-package.R
\docType{package}
\name{askoR-package}
\alias{askoR}
\alias{askoR-package}
\title{askoR: askoR - Differential Expresion Analysis using edgeR}
\description{
askoR is a pipeline for the analysis of gene expression data, using edgeR.
}
\keyword{internal}
